requirements.txt
pillow
SpeechRecognition==3.10.4
google-generativeai
pyttsx3